+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Refactor Aptean Respond API integration to give more control over form presentation
+- Add file field support to Aptean Respond forms
+- Custom sizes for input fields
 - Sitewide alert messages
 - Always show application forms front end component, even if the job ID is not imported